building: implement support for building with macOS universal2 bootloaders
Before appending data package to macOS bootloader, we need to ensure
that it is not signed. Typically, this is true for x86_64 bootloaders
(thin binaries or slices in universal2 fat binaries). However, arm64
bootloaders (thin binaries or slices in universal2 fat binaries) do
contain a signature due to stricter signature requirements on Apple
M1 Silicon platform. Therefore, if present, the signature needs to
be removed from the last slice (which, in universal2 binary, is arm64)
and then re-generated once the package data is appended.

However, it seems that when cross-compiling universal2 bootloader with
osxcross, we end up with both x86_64 and arm64 slice signed. While
the signature on the former should not have any impact, it does seem
to cause issues when trying to sign the final binary using the
real identity (codesign returns "invalid or unsupported format for
signature"). Therefore, we strip signatures from all slices, and
after data is appended, re-generate dummy signature on all slices.
These dummy signatures on the final binary can serve as integrity
validation in cases when users do not provide their own signature.

Add a new option in EXE(), called thin_arch, which converts the
universal2 fat binary bootloader to a thin binary of specified
architecture before appending the package data. Has no effect if
bootloader is already a thin binary.

For now, the option needs to be manually specified as an argument
to EXE() in the spec file.